Make it optional for parameter names to match named generators.
Previously, if parameter names were enabled at compile-time, all parameters were required to match a named generator. This made including parameter names, and using Lincheck with languages always enabling parameter names like Scala, impractical.

This instead makes it optional, such that if a named generator matches it is used but if none is found then we apply a default generator.

This also enables parameter names at compile time for tests in order to test this behavior.