Add preliminary support for inline methods in trace.
This commit adds preliminary support for recording Kotlin inline methods calls into trace.

It records void methods without arguments in class `<inline>`.

It is the only thing that is possible to implement with local transformation for now.

The problem is a marker variable (which marks inline call in bytecode) doesn't contain any class name or other information about the full method name.

Also, arguments of the inline method are not encoded at the call site. There ARE local variables that replace method arguments, but they are not distinguishable from local variables of inlined method, and the order of arguments is unknown.

Code needs to perform full Kotlin resolve by method name and analyze the class that contains the resolved method. It is not implemented (yet).

Other feature is not implemented is mapping line numbers via SMAP, so code locations can be bogus for now.