Lev serebryakov/tr bugfixes on kotlin tests (#752)
* Fix the logic of "transactions."

When a trace point with all its dependencies is larger than 1KiB, it can overflow per-thread buffer. In such case strategy flushes data to disk and re-runs the tracepoint save method.

Before this commit, it was implemented wrongly: its only flush & save.

Add logic to roll back the buffer and mark it before each object is written and reposition mark when the index cell is prepared (so, an object is written fully — index cell creation is done as the last step).

Limit string size on output.

Trim strings on input to 50 characters.